Add implementation for fetching featured and newest books
Implemented the fetchFeaturedBooks and fetchNewestBooks methods in HomeRepositoryImp class to fetch books from both local and remote data sources. If books are available in the local data source, they are returned, otherwise books are fetched from the remote data source. In case of any error, a Failure object is returned.